backend/requirements.txt
fastapi
uvicorn
duckdb
pandas
requests
python-dotenv
pydantic
streamlit
